\name{fusedest-package}
\alias{fusedest-package}
\docType{package}
\title{
Block splitting algorithms for estimation with fused penalty functions
}
\description{
It provides functions for carrying out block splitting algorithms for fused penalty estimation.
}
\details{
\tabular{ll}{
Package: \tab fusedest\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2019-07-17\cr
License: \tab GPL-2\cr
}
}
\author{
Tso-Jung Yen

Maintainer: Tso-Jung Yen <tjyen@stat.sinica.edu.tw>
}
\references{
Yen, T.-J. (2019). Solving fused penalty estimation problems via block splitting algorithms.
}
